Basic webapp project using django framework.
It my old basic project in webapp. In this project i working with django function based view. url mapping, templates and other basic of django framework.
And others in this project i create a static file for css, images, scss etc.
Create templates for html files and create a app for easy to work.